+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.4.0
+
+* added hourly visitor statistics, listing events, os and browser versions
+* added filter options
+* updated return types with new and modified fields
+
 ## 1.3.5
 
 * fixed domain method returning an object instead of an array